Handle unicode domains
The filesystem encoding reported by the docker image is ascii, which
triggered Twisted to convert unicode arguments to getProcessValue to
ascii as well.  Now does an explicit encoding as UTF-8, as we're not
referring to filesystem path.  Should getProcessValue throw an
exception, the domain that triggered it is removed from the queue.

A method to flush the running queue has been added.